Removing posted_in from category analysis
The initial idea of including both the comment body and the posted_in
text when doing the category extraction was that we wanted to keep the
context for where the comment was made when doing the sentiment
anlaysis. However the result is that we add a bit too much noise in the
final gauges. Therefore I believe this simplification makes better data.

Also adding versioning logic.